# builder-web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DST Builder",
    "description": "DST Builder 最小生成闭环 API（SPEC-DB-001 §11）",
    "version": "0.3.5"
  },
  "paths": {
    "/api/projects": {
      "post": {
        "summary": "Create Project",
        "description": "创建项目库；目录已是 Builder 项目时幂等打开（opened_existing=True，200）。",
        "operationId": "create_project_api_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectStateResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectStateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/current/draft": {
      "patch": {
        "summary": "Patch Current Draft",
        "description": "带 base_updated_at 乐观并发保存草稿；返回字段级诊断。",
        "operationId": "patch_current_draft_api_projects_current_draft_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectStateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets": {
      "post": {
        "summary": "Create Asset",
        "description": "纳入并哈希一个模板资产（§11）：复制 → 校验 → 内容寻址落盘 → 入库。",
        "operationId": "create_asset_api_assets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetIntakeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets/{asset_id}/inspect": {
      "post": {
        "summary": "Inspect Asset",
        "description": "用匹配版本 CAD 读取可用布局（§11）。\n\n生产默认惰性装配 CoreConsoleDrawingBuilder（Task 4\"端口未接线固定\n501\"已由真实执行器接线取代）；CAD 版本未配置或不可用时仍以\nCAD_VERSION_UNAVAILABLE（501）呈现，绝不伪造布局列表。",
        "operationId": "inspect_asset_api_assets__asset_id__inspect_post",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetInspectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetInspectionResponse"
                }
              }
            }
          },
          "501": {
            "description": "布局 inspection 端口未接线或匹配版本 CAD 不可用（CAD_VERSION_UNAVAILABLE）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayloadModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/plans/{plan_id}/confirm": {
      "post": {
        "summary": "Confirm Plan",
        "description": "用户显式确认计划（绝不自动确认）；确认后构建输入冻结。",
        "operationId": "confirm_plan_api_plans__plan_id__confirm_post",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanConfirmationResponse"
                }
              }
            }
          },
          "404": {
            "description": "计划不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayloadModel"
                }
              }
            }
          },
          "409": {
            "description": "PLAN_STALE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayloadModel"
                }
              }
            }
          },
          "422": {
            "description": "存在阻断诊断",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayloadModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/builds": {
      "post": {
        "summary": "Start Build",
        "description": "基于已确认计划创建 build/attempt 并在线程执行器中启动构建。",
        "operationId": "start_build_api_builds_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "计划不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayloadModel"
                }
              }
            }
          },
          "409": {
            "description": "PLAN_STALE / PACKAGE_TARGET_EXISTS / BUILD_ALREADY_RUNNING",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayloadModel"
                }
              }
            }
          },
          "422": {
            "description": "PLAN_NOT_CONFIRMED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayloadModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/builds/{build_id}/cancel": {
      "post": {
        "summary": "Cancel Build",
        "description": "请求安全取消（§6）：检查点处迁移 CANCELLED；PUBLISHING 拒绝。",
        "operationId": "cancel_build_api_builds__build_id__cancel_post",
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Build Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "构建不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayloadModel"
                }
              }
            }
          },
          "409": {
            "description": "PUBLISHING 或终止状态不响应取消（CANCEL_NOT_ACCEPTED）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayloadModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/builds/{build_id}": {
      "get": {
        "summary": "Get Build",
        "description": "读取状态、诊断和成果（§11）。",
        "operationId": "get_build_api_builds__build_id__get",
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Build Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "构建不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayloadModel"
                }
              }
            }
          },
          "500": {
            "description": "构建状态记录不完整",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayloadModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectCreateRequest": {
        "properties": {
          "project_root": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Root"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "discipline": {
            "type": "string",
            "title": "Discipline"
          },
          "output_path": {
            "type": "string",
            "title": "Output Path"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "stage",
          "discipline",
          "output_path"
        ],
        "title": "ProjectCreateRequest",
        "description": "POST /api/projects 请求。``project_root`` 仅在应用工厂未绑定根目录时必填。"
      },
...
      "ProjectStateResponse": {
        "properties": {
          "opened_existing": {
            "type": "boolean",
            "title": "Opened Existing",
            "default": false
          },
          "project": {
            "$ref": "#/components/schemas/ProjectModel"
          },
          "draft": {
            "$ref": "#/components/schemas/DraftFieldsModel"
          },
          "wizard_step": {
            "type": "integer",
            "title": "Wizard Step"
          },
          "focused_field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Focused Field"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "diagnostics": {
            "items": {
              "$ref": "#/components/schemas/DiagnosticModel"
            },
            "type": "array",
            "title": "Diagnostics"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project",
          "draft",
          "wizard_step",
          "focused_field",
          "updated_at",
          "diagnostics"
        ],
        "title": "ProjectStateResponse",
        "description": "GET /api/projects/current 与 PATCH /api/projects/current/draft 响应。"
      },
...
    }
  }
}